+++ FtxApi/Model/OrderModel.cs
@@ -24,14 +24,14 @@ namespace FtxApi.Model
         public string market { get; set; }
         public string type { get; set; }
         public string side { get; set; }
-        public decimal? price { get; set; }
-		public decimal? size { get; set; }
+        public double? price { get; set; }
+		public double? size { get; set; }
 		public string status { get; set; }
-		public decimal? filledSize { get; set; }
-		public decimal? remainingSize { get; set; }
+		public double? filledSize { get; set; }
+		public double? remainingSize { get; set; }
 		public bool reduceOnly { get; set; }
 		public string liquidation { get; set; }
-		public decimal? avgFillPrice { get; set; }
+		public double? avgFillPrice { get; set; }
 		public bool postOnly { get; set; }
 		public bool ioc { get; set; }
 		public string createdAt { get; set; }
